+++ pkg/storegateway/snappy_gob_codec.go
@@ -5,6 +5,7 @@ package storegateway
 import (
 	"bytes"
 	"encoding/gob"
+
 	"github.com/golang/snappy"
 	"github.com/pkg/errors"
 )